# src/docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My API",
    "description": "API documentation with swagger-autogen",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "tags": [
    {
      "name": "Users",
      "description": "User-related endpoints"
    }
  ],
  "paths": {
    "/lab/register": {
      "post": {
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/labCreationSchema"
              }
            }
          }
        }
      }
    },
    "/lab/config-setup": {
      "post": {
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/lab/list": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/lab/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/lab/by-user/{userId}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleCode",
            "in": "query",
            "description": "The ObjectId of the user to fetch labs for",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/lab/dependencies/graph": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/lab/by-alias/{alias}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/lab/detailed/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/lab/detailed/by-alias/{alias}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "alias",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/lab/{id}/exit_condition/{exit_condition_id}/update": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exit_condition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/conditionUpdateSchema"
              }
            }
          }
        }
      }
    },
    "/lab/{id}/update_level": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/levelStatusUpdateSchema"
              }
            }
          }
        }
      }
    },
    "/lab/{id}/assign-users": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/labRoleSchema"
              }
            }
          }
        }
      }
    },
    "/request/labs/create": {
      "post": {
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/labProposalSchema"
              }
            }
          }
        }
      }
    },
    "/request/labs/{id}/update": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/labProposalUpdateSchema"
              }
            }
          }
        }
      }
    },
    "/request/labs/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/request/labs": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/role": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/roleSchema"
              }
            }
          }
        }
      }
    },
    "/role/{code}": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/roleSchema"
              }
            }
          }
        }
      }
    },
    "/roles": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users": {
      "post": {
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userSchema"
              }
            }
          }
        }
      },
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/user-by-reference/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/{id}/roles": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/userRolesSchema"
              }
            }
          }
        }
      }
    },
    "/task-definitions/": {
      "post": {
        "description": "",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "short_description": {
                    "example": "any"
                  },
                  "description_url": {
                    "example": "any"
                  },
                  "version": {
                    "example": "any"
                  },
                  "associated_exit_conditions": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/task-definitions/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "labCreationSchema": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Dolphin watching Lab"
          },
          "alias": {
            "type": "string",
            "example": "DLPH-01"
          },
          "parent_lab_id": {
            "type": "string",
            "example": ""
          }
        },
        "required": [
          "name",
          "alias",
          "parent_lab_id"
        ]
      },
      "conditionUpdateSchema": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "example": 100
          },
          "comments": {
            "type": "string",
            "example": ""
          },
          "discussion_url": {
            "type": "string",
            "example": "http://google.com"
          }
        },
        "required": [
          "status",
          "comments",
          "discussion_url"
        ]
      },
      "levelStatusUpdateSchema": {
        "type": "object",
        "properties": {
          "level": {
            "type": "number",
            "example": 0
          },
          "state": {
            "type": "number",
            "example": 200
          }
        },
        "required": [
          "level",
          "state"
        ]
      },
      "roleSchema": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VLD"
          },
          "name": {
            "type": "string",
            "example": "Virtual Lab Developer"
          },
          "description": {
            "type": "string",
            "example": "Develops Virtual Labs"
          },
          "is_global": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "code",
          "name",
          "description",
          "is_global"
        ]
      },
      "userSchema": {
        "type": "object",
        "properties": {
          "reference_id": {
            "type": "string",
            "example": "12345"
          },
          "email": {
            "type": "string",
            "example": "test@test.nl"
          },
          "name": {
            "type": "string",
            "example": "Joe Doe"
          },
          "global_roles": {
            "type": "string",
            "example": "['VLD']"
          }
        },
        "required": [
          "reference_id",
          "email",
          "name",
          "global_roles"
        ]
      },
      "userRolesSchema": {
        "type": "object",
        "properties": {
          "roles": {
            "type": "string",
            "example": "['VLD']"
          }
        },
        "required": [
          "roles"
        ]
      },
      "labRoleSchema": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "user_id": {
              "type": "string",
              "example": "68417f10dbbf0245911c648c"
            },
            "role_codes": {
              "type": "string",
              "example": "[VLD]"
            }
          },
          "required": [
            "user_id",
            "role_codes"
          ]
        }
      },
      "labProposalSchema": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Dolphin watching Lab"
          },
          "alias": {
            "type": "string",
            "example": "DLPH-01"
          },
          "scope": {
            "type": "string",
            "example": "I need this lab for ..."
          },
          "timeplan": {
            "type": "string",
            "example": "It will be completed by the end of 2100"
          },
          "associated_users": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "string",
                  "example": "12345"
                },
                "role_codes": {
                  "type": "array",
                  "example": [
                    "GLU"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "lab_reference": {
            "type": "object",
            "properties": {
              "lab_id": {
                "type": "string",
                "example": "12345"
              },
              "lab_level": {
                "type": "number",
                "example": 0
              }
            }
          }
        },
        "required": [
          "title",
          "alias",
          "scope",
          "timeplan",
          "associated_users",
          "lab_reference"
        ]
      },
      "labProposalUpdateSchema": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "example": 10
          },
          "reviewer_user_id": {
            "type": "string",
            "example": "6860f128e861dfc672bcb368"
          },
          "comments": {
            "type": "string",
            "example": "test"
          }
        },
        "required": [
          "status",
          "reviewer_user_id",
          "comments"
        ]
      }
    }
  }
}